Update time when it's wildly incorrect (#39106)
Detect that the difference with the device's time is outside of a certain range (currently more than 5 minutes early or late), which means there's a loss of synchronization. If a synchronization loss is detected, then try to update the device's time. Avoid detecting synchronization loss too frequently (in case a device is misbehaving), currently we check again if more than an hour has passed since we last tried to correct it.