[E2E][JF] Fix jf-control-app README doc
All the instructions given in the README doc assumes that we are in the connectedhomeip repo, thus referencing the <TOP_LEVEL>.

The checkout_submodules.py script has to be executed from <TOP_LEVEL>.

Enhanced and fixed README documentation of jf-control (jfc) application.
Addition of 2 sections “Prepare for building” et “Checking if submodules are up to date”.
Remove the call to activate.sh during the build, as it doesn't work because of pigweed.