Start and commission test app before creating MTRDevice for it.
Fixes https://github.com/project-chip/connectedhomeip/issues/39477

What was happening is that we had an MTRDevice set up with a delegate already
while commissioning.  When the server advertised over DND-SD after AddNOC, the
MTRDevice canceled existing CASE setup attempts so it could try to set up a CASE
session to it, and this canceled the CASE setup attempt for
CommissioningComplete, so commissioning failed.

The fix is just to make sure commissioning is done before we set up our
MTRDevice in test049_CorrectTimeOnDevice.