Door lock controller: Add conformance markers
These are marked as EZ-target and EZ-initiator, but those
conditions aren't exposed in a way that alchemy understands.
These are both zigbee specific conditions, and they are marked
in the spec as [EZ-Initiator] and [EZ-Target]. This means they're
optional only for zigbee and disallowed for matter.

In later revisions, identify is removed from this device type.