TC-SC-4.3: Fix test for no TXT record
The spec is a BIT fuzzy in this area, but in my re-reading, a TXT
record is not specifically mandated, only certain keys in certain
situations (which would then require a TXT record).

If there are no mandatory TXT keys (which can happen for operational)
then the TXT record can be omitted and that's fine.