Add raw Zephyr SystemLayer implementation (noSelect) (#39505)
* [Zephyr] Add raw Zephyr SystemLayer implementation (noSelect)

The new Zephyr System Layer implementation does not use Zephyr's
BSD socket API (select). This lightweight option is suitable when
the Matter network layer interacts directly with the networking
stack, instead of going through Zephyr's NET L2 Layer. As a result,
it does not require sockets or the select API to function.

This implementation serves as an alternative to SystemLayerSelect,
and both can be used alongside Zephyr if needed.

Signed-off-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>
Co-authored-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>

* [nrfconnect] Add configuration for CHIP_USE_OPEN_THREAD_ENDPOINT

The new configuration allows enabling and disabling using
OpenThread endpoints as well as using Zephyr SystemLayer
implementation without Sockets.

Signed-off-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>
Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>

* [nrfconnect] Enable OpenThread endpoint in all-clusters-app

Enabled OpenThread endpoint instead of zephyr's socket in
nRFConnect all-clusters-app.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>

* Code review

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>

* Separate GenericPlatformManagerImpl_Zephyr.ipp files

Separated the Zephyr implementation of GenericPlatformManager
to version with and without using Select().

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>

---------

Signed-off-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>
Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>
Co-authored-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>