[NXP] Fix DnssdImplBr handling of startup and reset events (#39567)
This commit fixes the start-up logic of the platform mDNS
to return init success only after the wifi interface is
connected and the mDNS module is ready.
Also addresses the issue where after reset, in an multi-interface
scenario with WIFI and Thread enabled, the platform code will use
the Thread mDNS instead of the Wi-Fi one.

Signed-off-by: Marius Preda <marius.preda@nxp.com>