Fix build with chip_code_pre_generated_directory set
When setting chip_code_pre_generated_directory the build system is
expected to copy the pre-generated files using scripts/copyfiles.py.
It seems that the target directory is passed directly without using
rebase_path() to convert it to a relative path. With that, the command
fails with permission errors as the script attempts to write to /out.

Use rebase_path() to fix builds with chip_code_pre_generated_directory
set.

Signed-off-by: Stefan Agner <stefan@agner.ch>