[NXP][cmake][examples] Fix build issue when ICD feature disabled
The ICD manager is included in the Matter data interaction model and
this forces always the build of the ICDConfigurationData.cpp/h source
file pair, disregarding whether the device is an ICD or not. These
source files require at minimum that the ICD fast poll and ICD slow
poll intervals be defined (otherwise the build breaks), hence these
parameters must be defined at all times, even when the ICD feature is
disabled.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>