[NXP][cmake][examples] Don't stop for warnings when building Matter libraries in debug mode
In debug mode the Matter core libraries are generating quite a number
of build warnings and this frequently breaks the build if they are all
treated as errors. In order to avoid this issue, the
"treat_warnings_as_errors=false" build option will be used when
building the Matter SDK in debug mode.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>