Add a null-check on the return value of function getXTransitionStateByIndex
The function getXTransitionStateByIndex may return nullptr if index is invalid.
In function updateXYCommand, the return value of getXTransitionStateByIndex is directly propagated to function computeNewColor16uValue, where it is dereferenced without any validation.
I think it would be better to add a check on the return value of getXTransitionStateByIndex, as all other callsites have such checks.