Avoid editable install of py_matter_idl in build env (#39826)
* Avoid editable install of py_matter_idl in build env

The editable install creates a problem for the way the configure script creates
the build env, because it assumes that it can create a self-contained, reusable
build env that keeps working even if the original source directory is removed
or moved to a different location. This is no longer true when the build env
contains effectively a "symlink" into the source tree.

This change instead modifies codegen.py and codegen_paths.py to directly find
py_matter_idl relative to the script directory and import them from there,
without relying on the editable install. The python-path pip package is used to
do this (essentially it's just a small utility for resolving a path and
prepending it to sys.path, allowing the code in the script itself to remain
readable and straightforward.)

* Avoid empty path in error message

* Include py_matter_idl dependencies in requirements.build.txt

* Ensure CDPATH doesn't affect configure script