Fix reading ExtendedAddress from otbr-agent (#39723)
* Fix reading ExtendedAddress from otbr-agent

The property is defined to not emit signals on changes, which means
the generated caching proxy will never have a value. Instead, we must
directly read the value with the Get function and cache it ourselves.

Fixes #39722

* Review fixes

- added missing error return
- use CHIP_ERROR_KEY_NOT_FOUND instead of CHIP_ERROR_INTERNAL, following
  _GetThreadProvision implementation.

* restyled

* review comments

use VerifyOrReturnError for neatness.

* Remove unused DBusOpenthread properties

Since Linux platform uses a caching proxy to read OpenThread properties
over dbus, it will not provide values for properties annotated as
EmitsChangedSignal=false, so it is misleading and error-prone to
generate get function accessors for those. These properties were
therefore removed from generation. They werent being used anyway.

* Revert "Remove unused DBusOpenthread properties"

This reverts commit 923a8da01a294e82d98a7534d63f17b5a341b43a.

* make referenced otbr dbus props writable only

They are being written to, so the caching proxy isnt an issue
but we still need the properties in the generated interface.