Do some size optimization on AttributeValueEncoder. (#39822)
* Do some size optimization on AttributeValueEncoder.

This tries to avoid separate template instantiations of Encode methods for
various integer types and bitmap types, instead funneling all of those into the
pair of instantiations for uint64_t and int64_t.

We had some existing machinery for the unsigned integer types that this
replaces; the bitmap and signed integer bits are new and mean that callers don't
have to manually call Raw() on their bitmaps to get the space savings.

* Address review comments.