[NXP][wifi] Improving wifi port state machine in case of wrong wifi credential given during the commissioning process
This change aims to  address Wi-Fi connection failures during Matter commissioning.
For instance, if an incorrect SSID or password is provided, we new  mechanism to stop the Wi-Fi driver from attempting to connect has been put in place to allow  a new connection attempt later.
If the failSafeTimer expires while the system is still in the process of connecting, a disconnect event is scheduled.
This ensures that, once disconnected, the wlan_remove_network function is called as part of the RevertConfiguration process.

Signed-off-by: Gatien Chapon <gatien.chapon@nxp.com>