[NXP][CMake][MCXW7x] Fix build issues when enabling "factory-data" and "OTA" (#39853)
* [NXP][cmake][examples] Fix build issue when ICD feature disabled

The ICD manager is included in the Matter data interaction model and
this forces always the build of the ICDConfigurationData.cpp/h source
file pair, disregarding whether the device is an ICD or not. These
source files require at minimum that the ICD fast poll and ICD slow
poll intervals be defined (otherwise the build breaks), hence these
parameters must be defined at all times, even when the ICD feature is
disabled.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][examples][mcxw7x] Disable ICD feature by default

ICD behavior makes sense only for MTDs (Minimal Thread Devices)
therefore it should be disabled by default for any other type of
devices.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][examples][mcxw7x] Enable persistent subscriptions for MTDs

Subscription persistence should be enabled for all MTDs (Minimal
Thread Devices) generated with CMake in order to align them with the
applications generated via GN.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][platform][mcxw71] Build Matter libs with "-Og" optimization rather than "-O0" in debug mode

The MCXW71 platform internal flash size does not afford the use of the
"-O0" optimization level in debug mode for Matter core libraries and,
therefore, an optimization level which generates a reduced size binary
image is more appropriate. "-Og" is a debug optimization level which
generates reduced size binary images while still providing a reasonable
debuging experience.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][examples] Don't stop for warnings when building Matter libraries in debug mode

In debug mode the Matter core libraries are generating quite a number
of build warnings and this frequently breaks the build if they are all
treated as errors. In order to avoid this issue, the
"treat_warnings_as_errors=false" build option will be used when
building the Matter SDK in debug mode.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

---------

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>