DM XMLs: Add script steps to cleanup old spec revisions
As discussed on slack, there are some older spec revisions that
do not have specific branches, but do have errata that affect
the data model scrapes. The DM team has opted to change these
in the SDK code rather than attempting to force a new branch
on the spec for the sake of small typos.

There is a possibility that the SIT/LIT problem can be addressed
in alchemy, but for expedience I am fixing it here and will remove
once alchemy is updated.