[app] Fix crash when session is replaced before priming reports
...are sent.

When session is replaced with a new one after Subscribe
Request but before all priming reports have been sent,
the exchange context's session is not switched to the
new session because of its kStayAtOldSession policy.

The Read Handler code was not expecting that and tried
to dereference a released session handle.

Related to #40190

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>