Fix for chip-cert-bins Dockerfile: Add libpcsclite1 and pcscd  (#40153)
* Fix for chip-cert-bins Dockerfile:
A couple of months ago, libpcsclite1 and pcscd have been added (see PR 38599).
Those packages are needed to access to an USB NFC Reader from chip-cert-bins
 container.
Unfortunately, they have been added to the Stage 1 of Dockerfile, used to
 build the container. Consequently, they are not available at runtime.

The current PR is moving them to the Stage 3, by this way they will be
available at runtime.

I also took this opportunity to execute the command "pcscd --disable-polkit"
which is needed to access a PCSC reader from a docker container.

Change-Id: I5147a9bb5c7ac8c5bf0f3591868f0943ecba3221

* Restyled by whitespace

* Removed --disable-polkit pcscd command.

Change-Id: I0895947e26209b88b9068d7e705e7d5b50960ebb

* Added libpcsclite-dev for build stage.

Change-Id: I385f766eae0200e20bad6d3c48722346a6090338

---------

Co-authored-by: Restyled.io <commits@restyled.io>