DM XMLs: Add code for fixing known issues with prior spec revisions (#40209)
* DM XMLs: Add script steps to cleanup old spec revisions

As discussed on slack, there are some older spec revisions that
do not have specific branches, but do have errata that affect
the data model scrapes. The DM team has opted to change these
in the SDK code rather than attempting to force a new branch
on the spec for the sake of small typos.

There is a possibility that the SIT/LIT problem can be addressed
in alchemy, but for expedience I am fixing it here and will remove
once alchemy is updated.

* Update 1.3 DM files

* Update 1.4 DM files

* Update 1.4.1 DM files

* Update master DM XMLs

* Minor correction to doc file to omit master

* Add scraper script to CI

It's not a full scrape yet, but it's getting closer and at least
now we're enforcing the script works.

* Restyled by isort

* review comments: Appease the bot

* DNS - using this to show CI failure

* Add lxml to package checker

* Revert "DNS - using this to show CI failure"

This reverts commit e0e4dc41d93e06978cd6937e2d0939a3b06bcd10.

---------

Co-authored-by: Restyled.io <commits@restyled.io>