Add data model XML files for 1.2 spec (#39535)
* add files

* Add 1.2 to python whl

* Add prebuilt for 1.2

* Add new 1.2 files to python build

* huge hack - add base device clusters

Pre-1.3, the tables for these were divided into zigbee and matter.
Alchemy doesn't know about the older tables.

Given that this spec is stable now, I'm doing this manually from
the later revisions, since the revisions for this device type
has not changed.

We could write this into alchemy, but the effort / benefit ratio
there seems low.

* Door lock controller: Add conformance markers

These are marked as EZ-target and EZ-initiator, but those
conditions aren't exposed in a way that alchemy understands.
These are both zigbee specific conditions, and they are marked
in the spec as [EZ-Initiator] and [EZ-Target]. This means they're
optional only for zigbee and disallowed for matter.

In later revisions, identify is removed from this device type.

* Add 1.2 to device type unit tests

* Add 1.2 to selection and cluster tests, fix selection

* address review comments

* Add missing OTA cluster files to 1.2

These are included weirdly in the 1.2 spec. Copy from 1.3. This is
a one-off, so I'm not fixing the scraper to deal with a spec
revision that's nearly sunset.

* manually add OTA clusters to 1.2 cluster id json

* Add new files to python build gn

* Update cluster IDs for OTA device types

Even though these clusters were missing anyway, that wasn't the
only problem - the device types used the IDs of the device
types rather than the clusters. This is all very fragile.
We need more spec checks.

* Update doc files for 1.2

* Restyled by isort

---------

Co-authored-by: Restyled.io <commits@restyled.io>