Create super-class OperationalDatasetView out of OperationalDataset
OperationalDatasetView does not copy or hold a buffer for the underlying
dataset, and provides all the read-only methods that were present on
OperationalDataset.  OperationalDataset has its own buffer and copies the
dataset as before.

Also simplify some of the mechanics of updating values in the dataset, by
replacing MakeRoom with InsertOrReplace that directly takes the required size
of the value and handles updating both the tag and length of the TLV item as
well as of the overall dataset, so the caller only has to fill in the value.