Update quality properties in matter_idl_types.py
Change the various quality properties in matter_idl_types.py to use the `in` operator instead of `&`, and declare each to return a `bool` value.

Currently these properties use the `&` operator as a set intersection with the enum flag values.  This leads to these methods unexpectedly returning instances of the quality flags.