[NXP] Constructing trusty-matter object using a global variable
TrustyMatter calls its own constructor when instantiated, which
establishes a communication channel with the Trusty OS.
However, the number of communication channels that Trusty OS can
establish is limited. Therefore, it is not appropriate to
instantiate TrustyMatter as a class member.
This patch modify TrustyMatter as a global variable to instantiate
to avoid repeated construction.

Change-Id: Icd898d57b603db6c7be3727bcb19e047fa25e714
Signed-off-by: Maximus <maximus.sun@nxp.com>