Make FindCommissioneeDevice more reliable. (#40759)
* Make FindCommissioneeDevice more reliable.

Right now, whether FindCommissioneeDevice works or not depends on whether the
initial "commmissioning ID" matches the final node ID assigned to the node and
on whether the call to FindCommissioneeDevice happens before or after AddNOC.

This means, in particular, that StopPairing is not very reliable, because it can
race the AddNOC bits and then you don't know what ID to pass to StopPairing.

This change allows either the original or the updated ID to be used to identify
the commissionee device.

* Address review comment.