Notify errors if StopPairing is called before PASE is set up. (#40766)
If StopPairing happens after PASE is set up and we are doing our state machine,
we send out a notification that the relevant commissioning step failed.

But a StopPairing during PASE setup would just silently stop things and not
deliver any notifications.

This fix makes sure that in that situation we deliver a "pairing failed"
notification (which is all we can do at that level).