TC-SC-7.1: Add PICS per CCB 4232 (#40296)
As the CCB points out, this test is currently required for all
commissionable devices, but is actually only applicable for
devices that include a pairing code since the test requires it.
This is MOST of the commissionable devices, but doesn't apply to
devices that are field-upgrading to Matter.
The correct PICS as discussed in the TSG is the MANUAL PC PICS.
Devices may opt to provide the QR code, but the test is gated
on the manual PICS since all devices with a QR code will have a
manual code (mandatory), but the QR code is technically optional.