[NXP][platform][mcxw71] Add CMake build support (#40925)
* [NXP][examples] Add CMake build support for "lighting-app" and "contact-sensor-app"

Signed-off-by: Dina Benamar <dina.benamarelmaaroufi@nxp.com>
Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][examples][mcxw7x] Fix low-power app flavors not operating in low-power mode

The low-power build configurations of the Matter apps for MCXW7x
platforms were not operating in low-power mode due to the fact that
an incorrect configuration of build flags was being used.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][examples] Fix incorect use of CONFIG_NET_L2_OPENTHREAD build option

In several Kconfig build configuration files CONFIG_CHIP_NET_L2_OPENTHREAD
was used by mistake instead of CONFIG_NET_L2_OPENTHREAD.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][examples][mcxw71] Fix missing dependency on "openthread_cli" when Matter CLI enabled

The NXP platform library is missing a dependency on the "openthread_cli"
module when Matter CLI is enabled and this is breaking the build of
Matter apps that are providing Matter CLI.

Signed-off-by: Dina Benamar <dina.benamarelmaaroufi@nxp.com>
Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][examples] Add CMake build support for "lock-app"

Signed-off-by: Oliver Fan <oliver.fan@nxp.com>
Signed-off-by: Dina Benamar <dina.benamarelmaaroufi@nxp.com>
Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][examples] Allow the user build config file to take precedence over the default application build config

The user was unable to change the default application build settings
because the build config file received via the CONF_FILE_NAME CMake
variable was incapable of overwriting the default application build
settings. This update reverses the order of assembly of build settings
and allows the user build config file to take precedence over the
default application build config.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][platform][mcxw7x] Don't use header paths relative to the CHIP root

The CHIP project policy is to use header paths which are relative
to the source subfolder ("src"). Nobody is using header paths which
are relative to the CHIP project root directory.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][platform][mcxw71] Increase SHA256 context size for MbedTLS SSSAPI port

The standard NXP MbedTLS SSSAPI based implementation requires a SHA256
context size of at least 136 bytes.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][platform][mcxw7x] Update build configuration to build "MbedTLS" using CMake

In the context where the Matter apps are built using 2 different
build systems - a CMake build controlling a "gn" build - up until
this point the "MbedTLS" component was compiled using "gn" and
included in the Matter platform support.

This upgrade restructures the build so that the "MbedTLS" is compiled
by the outer CMake build process instead for enhanced flexibility.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>
Reviewed-by: Dina Bemanar El Maaroufi <dina.benamarelmaaroufi@nxp.com>
Reviewed-by: Gatien Chapon <gatien.chapon@nxp.com>

* [NXP][cmake][build_examples] Fix Matter apps build issues using CMake

When building the Matter apps for the MCXW71 platform using CMake
specific arguments are required.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][cmake][build_examples][mcwx71] Enable building low-power app flavors

Added some updates to allow building low-power app flavors for the
MCXW71 platform using CMake and "build_examples.py".

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

* [NXP][workflow][mcxw71] Test CMake builds instead of GN builds

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>

---------

Signed-off-by: Dina Benamar <dina.benamarelmaaroufi@nxp.com>
Signed-off-by: Marian Chereji <marian.chereji@nxp.com>
Signed-off-by: Oliver Fan <oliver.fan@nxp.com>
Co-authored-by: Dina Benamar <dina.benamarelmaaroufi@nxp.com>
Co-authored-by: Oliver Fan <oliver.fan@nxp.com>