Add a DevicePairingDelegate notification for a commissioning stage starting. (#41000)
* Add a DevicePairingDelegate notification for a commissioning stage starting.

We had notifications for a stage ending, but not for a stage starting.  Because
you can't tell which stage will start next from an end notification (without
replicating the commissioner state machine), and because some stages can take
quite a while, commissioners want to be notified when stages start, not just
when they end.

* Address review comment.