Add support for requesting network credentials without a scan.
Before this change, CHIPDeviceController only supported two modes of operation:
providing network credentials up front to commisisoning, or doing a network scan
and then asking the delegate for credentials.

This change adds a third mode: once we have determined which network
technologies the commissionee supports, requesting credentials for those
specific network technologies without doing a scan.  This avoid the need for
commissionees to find unnecessary credentials up front while still allowing them
to provide the credentials the commissionee can actually use.

The chip-tool changes to make use of this are not ideal, but do make it possible
to exercise the new codepaths and ensure they work.