[nrfconnect] Remove dependency to DK in Matter configuration
In Matter configuration there must not be anything related to
DK. All DK-related Kconfigs must be defined in the application.

The CHIP_SPI_NOR and CHIP_QSPI_NOR configs are no longer needed
and we can use NORDIC_QSPI_NOR and SPI_NOR instead.

According to the Zephyr guidelines, do not write "Enable"
for boolean prompts.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>