chip-tool should advertise on loopback interface only when running CI tests. (#41315)
When running in interactive mode (which is what CI tests use for running
chip-tool based tests), chip-tool advertises operationally.  It was advertising
on all interfaces, and it uses the same hostname as the server app, so when it
tried to send commands to the server app it could end up with a non-loopback
IP.  And this sometimes failed in CI with "Network is unreachable".

This change:

* Adds support to CHIPDeviceControllerFactory init for setting an interface id
  to use, like Server init already has.
* Adds support to chip-tool to take an "interface-id" optional argument in
  interactive modes and propagate it through to controller factory init.
* Passes "--interface-id -1" as an argument to chip-tool from the CI test
  harness.