Darwin: Expose MTRSetupPayload subPayloads and concatenated properties (#41450)
* Darwin: Expose MTRSetupPayload subPayloads and concatenated properties

Re-generate the qrCodeString from the actual sub-payloads on demand. This means
sub-payloads can be mutated without requiring special handling or API.

Attempting to mutate a concatenated payload object as if it wasn't concatenated
drops the subsequent payloads, i.e. legacy code that is unaware of concatenated
codes (but uses MTRSetupPayload in a mutable way) will simply work on the first
payload.

Expand the tests for concatenated codes.

* restyle

* Update src/darwin/Framework/CHIP/MTRSetupPayload.mm

Co-authored-by: Boris Zbarsky <bzbarsky@apple.com>

* Update src/darwin/Framework/CHIP/MTRSetupPayload.h

Co-authored-by: terafin <jwood@me.com>

* Update src/darwin/Framework/CHIP/MTRSetupPayload.h

Co-authored-by: terafin <jwood@me.com>

* Apply suggestion from @woody-apple

* Apply suggestion from @woody-apple

---------

Co-authored-by: Justin Wood <woody@apple.com>
Co-authored-by: Boris Zbarsky <bzbarsky@apple.com>
Co-authored-by: terafin <jwood@me.com>