Fix parsing InstanceName in UDCServer when missing nullptr
Zero-initialize the udcPayload buffer for OnMessageReceived and
pass back the processed udcPayloadLength when reading the
IdentificationDeclaration of the UDC packet.
This fix addresses issue #41748.

[Problem]
On receipt of udcPayload where InstanceName is missing nullptr, it is possible
the instanceName can include an extra random byte due to a non-zero initialized
buffer and passing back of the udcPayload with the incorrect length.

* Restyled changes