Python wheels: Create one wheel to rule them all
We are working towards launching a python support module on PyPI
to support local testing for developers who do not work day-to-day
on the SDK and would prefer to just pip install a package.

This package contains a single wheel that encompases the following
individual wheel files from the SDK
- matter-core
- matter-clusters
- matter-repl
- matter-idl
- matter-yamltest
- matter-testing

matter-jupiter will be added in a subsequent PR.

The three core matter wheel files are currently being compiled
by a secondary script. I opted to keep this in place for this PR
to avoid complicating changes, but we should be able to remove
the duplication in a follow up and compile the -core, -clusters
and -repl wheel files through pigweed for simpliciy.

This PR aims just to match the current python environment setup
and therefore still requires that the test scripts themselves
be provided. This is another follow up, as is integration of the
mdns into the testing framework and inclusion of the support files
like the certificates.

A CI test will be added to this PR in a subsequent commit.

Follow up summary for this PR:
- CI to 1) create raw venv 2) install from combined wheel
  3) run at least one tests (that uses data_model)

Follow up summary:
- CI job to manual push to test PyPI
- add jupiter
- change -core -clusters -repl to use pw build files
- add mdns files to matter-testing
- add an additional wheel to bundle the tests and the certs
- examine swapping the CI over to pip install