Update test_metadata.yaml
Updated as tests CADMIN_1_3 and CADMIN_1_4 are both contained in the same test module TC_CADMIN_1_3_4.py instead of seperate files. Was not originally sure if the name was based off the tests or the test file name. Appears according to the other item names listed its based off the test file names.