DM XMLs: Remove master directory (#42024)
* DM XMLs: Remove master directory

The intention was that this directory would be updated regularly
and reflect the current master with all includes. That's not been
happening and this directory is now deeply out of date. It also
doesn't necessarily make sense to have a "master" data model when
we have multiple ongoing releases in the spec with different includes.

This directory of files is out of date and misleading. Let's remove
it for now, and if we decide later that we want to have a dependabot
supported pull from master with some set of includes, we can bring
it back at that point. But for now, let's get rid of it and avoid
confusion.

* also remove from CI checks