Make it easier to configure CodegenIntergration clusters from app code
The main change here is to make CodegenDataModelProvider have an Init() method
that's distinct from Startup() (which is called from Server::Init()). After
Init(), codegen-integration clusters will have been created but not started,
making it straightforward for application code to obtain references to cluster
instances and doing things like setting a delegate.

Linux AppMain.h now exposes ChipLinuxDefaultServerInitParams() which creates
the Server dependencies including the initialized (but not started) data model
provider.

Also make the resources held by CommonCaseDeviceServerInitParams non-static
members; this makes the class behave sanely if more than one instance is
created, e.g. in a test. In all in-tree usages, the
CommonCaseDeviceServerInitParams instance itself is already static anyway, so
this doesn't change the lifetime of the resources.