Add delays for value propagation in TC_I_2_4.py
Added delays to allow value propagation for non Matter SDK implementations.

Background: The Matter SDK emits new values to subscriptions immediately as soon as one new value exists. This can lead to multiple subscription messages when the change of a value potentially affects others too very fast afterwards.

matter.js uses a slight collection delay of 50ms to collect values to try to send out as least subscription messages as possible and in normal use cases this delay does not hurt anyone.
The spec is also not really stating the expected behavior here.

This test changes the IdentifyTime value very hast after each other (usually 5-10ms) and expected that both values were reported. For the matter,js "collection approach" this was too fast and so the initial value was never reported. 

This slight delay here ensures that the behavior is as the test expects.

### Testing

This PR adjusts a test