TestTCP: Make HandleConnCloseTest[46] work reliably on Darwin
Wait until after HandleConnectionReceived() is called before closing the client
connection, otherwise we can't be sure the connection actually gets accepted at
the application level (as opposed to in the kernel) on the server side.

Also add some logging to the callbacks, and check that the callbacks are
actually being called with the expected handles. Remove some unused test code.