TCP: Remove the fallback for GetPeerAddress
The updated tests show this is no longer needed. I encountered no test failures
in 1000 repetitions on either Linux or Darwin. Certain test scenarios have a
chance of logging a "Failure accepting incoming connection" error, but these
are cases where the client-side socket is being closed as soon as it's
connected. Particularly on Darwin this may then cause the GetPeerAddress call
to fail (Linux doesn't seem to exhibit this behavior.)