TCP: Make HandleConnCloseTest[46] work reliably on Darwin and related fixes (#42136)
* TestTCP: Retry with different ports on EADDRINUSE

* TestTCP: Make HandleConnCloseTest[46] work reliably on Darwin

Wait until after HandleConnectionReceived() is called before closing the client
connection, otherwise we can't be sure the connection actually gets accepted at
the application level (as opposed to in the kernel) on the server side.

Also add some logging to the callbacks, and check that the callbacks are
actually being called with the expected handles. Remove some unused test code.

* TCPConnect base implementation should not return CHIP_NO_ERROR

* Darwin: Detect saLen == 0 from accept() and ignore the connection

* TCPTest: Add test for "late" failure (in TCP.DoHandleIncomingConnection)

* TCP: Remove the fallback for GetPeerAddress

The updated tests show this is no longer needed. I encountered no test failures
in 1000 repetitions on either Linux or Darwin. Certain test scenarios have a
chance of logging a "Failure accepting incoming connection" error, but these
are cases where the client-side socket is being closed as soon as it's
connected. Particularly on Darwin this may then cause the GetPeerAddress call
to fail (Linux doesn't seem to exhibit this behavior.)

* From review: make test macro docs less verbose