[NXP] Fix possible double free of the resolve context
This commit fixes a rare case where the resolve context is being freed a
second time in a Dispatch function. To prevent this scenario a new function was
added, IsResolveElementValid() to check if the current resolve context is present
in the resolve list. If that is false, it means the context was freed by
NxpChipDnssdShutdown or NxpChipDnssdResolveNoLongerNeeded.

Signed-off-by: Marius Preda <marius.preda@nxp.com>