Fix validation issues in MTRThreadOperationalDataset
When initializing from an NSData, the MTRThreadOperationalDataset code
currently ignores errors from any of the Get*() methods it calls on the
underlying chip::Thread::OperationalDataset. This means that if the dataset
contains any field (for example a MasterKey) of invalid length, the MTR object
has its corresponding property populated with garbage (from the uninitialized
stack buffer).

_populateCppOperationalDataset also did not check for errors, but this call
should not fail in practice since we pre-check the parameter lengths when
initializing with individual properties.

The PanID handling is also broken; it's documented as being a 2 byte buffer in
host byte order (little endian on Apple platforms), but the implementation
byteswaps it before passing it to chip::Thread::OperationalDataset. However,
that method takes a uint16 (NOT a buffer) and has always correctly handled
encoding and decoding to the underlying TLV in network byte order. The byte
swapping was added in #8562, but it's unclear why this was thought to be
necessary.